Fixes #7. changed csv file to append if exists, times now shown to all decimal places (a lot!), added quality column to csv which gives optimal/actual, added repetitions option - so batch times can be averaged over number of attempts

+++ src/p4_controller.py
@@ -91,9 +91,14 @@ class SimController(object):
             self.readConfig()
             self.gen = self.stepGenerator(self.cfg["START"], self.cfg["GOAL"])
         elif self.cfg["BATCH"] is not None:
-            self.runBatch(self.cfg["BATCH"][0],self.cfg["BATCH"][1])
-            print("\nBatch process completed. Results written to "+ self.cfg["BATCH"][1]+".\n")
-            raise SystemExit()
+            try:
+                self.runBatch(*self.cfg["BATCH"])
+                print("\nBatch process completed. Results written to "+ self.cfg["BATCH"][1]+".\n")
+            except Exception as e:
+                print("\nAn error has occurred. Batch results may be incomplete.")
+                print e
+            finally:
+                raise SystemExit()
         else:
             try:
                 self.setStart(ast.literal_eval(self.cfg.get("START")))
@@ -383,21 +388,22 @@ class SimController(object):
 
     def hdlStop(self):
         """Button handler. Displays totals."""
-        if self.cfg.get("AUTO"):
-            message =(str(self.pathcost) + ";" + \
-                  str(self.pathsteps) + ";" + \
-                  '{0:.5g}'.format(self.pathtime) + ";" + \
-                  '{0:.5g}'.format(self.timeremaining))
-            return message
+        if isinstance((self.pathcost),int):
+            totalcost = str(self.pathcost)
         else:
-            if isinstance((self.pathcost),int):
-                totalcost = str(self.pathcost)
-            else:
-                totalcost = '{0:.2f}'.format(self.pathcost)
+            totalcost = '{0:.2f}'.format(self.pathcost)
+        if self.cfg.get("AUTO"):
+            message =(totalcost + ";" + \
+                str(self.pathsteps) + ";" + \
+                str(self.pathtime) + ";" + str(self.timeremaining))
+            
+            return message
+
+        else:
             message = "Total Cost : " + totalcost + \
                       " | Total Steps : " + str(self.pathsteps) + \
-                      " | Time Remaining : " + '{0:.5f}'.format(self.timeremaining) + \
-                      " | Total Time : " + '{0:.5f}'.format(self.pathtime)
+                      " | Time Remaining : " + str(self.timeremaining) + \
+                      " | Total Time : " + str(self.pathtime)
 
         self.updateStatus(message)
 
@@ -461,7 +467,7 @@ class SimController(object):
                         " | Steps : " + str(self.pathsteps)
                     if self.cfg.get("DEADLINE"):
                         message += " | Time remaining: " + \
-                                   '{0:.5f}'.format(self.timeremaining)
+                            str(self.timeremaining)
                     self.updateStatus(message)
                     sleep(self.cfg.get("SPEED"))  # delay, if any
                     
@@ -581,36 +587,53 @@ class SimController(object):
         except: #we don't care why it failed
             self.updateStatus("Failed to load script.py")
             
-    def runBatch(self, infile, outfile):
-        # assumes MAP_FILE, AGENT_FILE, BATCH[0], BATCH[1] set in self.cfg
+    def runBatch(self, infile, outfile, reps = 1):
+        # assumes MAP_FILE, AGENT_FILE set in self.cfg
         # initialise map and agent
         print ("\nRunning batch...")
+        times = []
+        reps = int(reps)
         self.processMap()
         self.initAgent()
+        self.processPrefs() 
         # open scenario file and read into problems list
         scenario = open(infile)
         problems = [line.strip().split() for line in scenario if len(line) > 20]
         scenario.close
 
-        # Open csv file and processs each problem and dump results in csv file
-        with open(outfile, 'wb') as csvfile:
+        # If csv file doesn't exist, create and write header, then close
+        if not os.path.isfile(outfile):
+            with open(outfile, 'wb') as csvfile:
+                fcsv = csv.writer(csvfile, delimiter=',',
+                                        quotechar='|', quoting=csv.QUOTE_MINIMAL)
+                fcsv.writerow(['agent', 'map', 'startx', 'starty', 'goalx', 'goaly', 'optimum', 'actual', 'steps', 'time_taken', 'quality'])
+        # Open existing csv file, process each problem and append results     
+        with open(outfile, 'ab') as csvfile:
             fcsv = csv.writer(csvfile, delimiter=',',
                                     quotechar='|', quoting=csv.QUOTE_MINIMAL)
-            fcsv.writerow(['agent', 'map', 'startx', 'starty', 'goalx', 'goaly', 'optimum', 'actual', 'steps', 'time_taken'])
-
             count = 1
             for problem in problems:
                 print "\r", count,  # output number of problems completed
                 count += 1
-                self.agent.reset()
                 skip, mappath, size1, size2, scol, srow, gcol, grow, optimum = problem
                 pathname, map = os.path.split(mappath)
                 self.cfg["START"] = (int(scol), int(srow))
                 self.cfg["GOAL"] = (int(gcol), int(grow))
-                self.resetVars()
-                output = self.search()
-                actual_cost, steps, time, inf = str.split(output, ';')
-                fcsv.writerow([self.cfg["AGENT_FILE"], map, str(scol), srow, gcol, grow, optimum, actual_cost, steps, time])   
+                
+                for i in xrange(reps):
+                    self.agent.reset()
+                    self.resetVars()
+                    output = self.search()
+                    actual_cost, steps, time, inf = str.split(output, ';')
+                    times.append(float(time))
+                    
+                time = sum(times)/reps    # calculate average
+                
+                try:
+                    quality = float(optimum)/float(actual_cost)
+                except ZeroDivisionError:
+                    quality = 0
+                fcsv.writerow([self.cfg["AGENT_FILE"], map, str(scol), srow, gcol, grow, optimum, actual_cost, steps, time, quality])   
 
             
 if __name__ == '__main__':
